+++ backend/canisters/cycles_dispenser/impl/src/guards.rs
@@ -9,9 +9,9 @@ pub fn caller_is_governance_principal() -> Result<(), String> {
 }
 
 pub fn caller_is_authorized_to_add_canister() -> Result<(), String> {
-    if read_state(|state| state.is_caller_authorized_to_add_canister()) {
+    if read_state(|state| state.is_caller_governance_principal() || state.is_caller_registry_canister()) {
         Ok(())
     } else {
-        Err("Caller is not authorized to add a canister".to_string())
+        Err("Caller is not authorized to add canisters".to_string())
     }
 }