+++ backend/canisters/cycles_dispenser/impl/src/lib.rs
@@ -5,7 +5,7 @@ use ic_ledger_types::{BlockIndex, Tokens};
 use ledger_utils::default_ledger_account;
 use serde::{Deserialize, Serialize};
 use std::cell::RefCell;
-use std::collections::{BTreeMap, BTreeSet, HashSet};
+use std::collections::{BTreeMap, HashSet};
 use types::{BuildVersion, CanisterId, Cycles, Milliseconds, TimestampMillis, Timestamped};
 use utils::env::Environment;
 
@@ -37,10 +37,8 @@ impl State {
         self.data.governance_principals.contains(&self.env.caller())
     }
 
-    pub fn is_caller_authorized_to_add_canister(&self) -> bool {
-        let caller = self.env.caller();
-        self.data.governance_principals.contains(&caller)
-            || self.data.canisters_directly_controlled_by_sns_root.contains(&caller)
+    pub fn is_caller_registry_canister(&self) -> bool {
+        self.env.caller() == self.data.registry_canister_id
     }
 
     pub fn metrics(&self) -> Metrics {
@@ -60,8 +58,11 @@ impl State {
             min_cycles_balance: self.data.min_cycles_balance,
             icp_burn_amount: self.data.icp_burn_amount,
             stable_memory_sizes: memory::memory_sizes(),
-            ledger_canister: self.data.ledger_canister,
-            cycles_minting_canister: self.data.cycles_minting_canister,
+            canister_ids: CanisterIds {
+                registry: self.data.registry_canister_id,
+                ledger: self.data.ledger_canister,
+                cmc: self.data.cycles_minting_canister,
+            },
         }
     }
 }
@@ -70,8 +71,8 @@ impl State {
 struct Data {
     pub governance_principals: HashSet<Principal>,
     pub canisters: Canisters,
-    #[serde(default)]
-    pub canisters_directly_controlled_by_sns_root: BTreeSet<CanisterId>,
+    #[serde(default = "CanisterId::anonymous")]
+    pub registry_canister_id: CanisterId,
     pub sns_root_canister: Option<CanisterId>,
     pub max_top_up_amount: Cycles,
     pub min_interval: Milliseconds,
@@ -89,6 +90,7 @@ impl Data {
     pub fn new(
         governance_principals: Vec<Principal>,
         canisters: Vec<CanisterId>,
+        registry_canister_id: CanisterId,
         max_top_up_amount: Cycles,
         min_interval: Milliseconds,
         min_cycles_balance: Cycles,
@@ -101,7 +103,7 @@ impl Data {
         Data {
             governance_principals: governance_principals.into_iter().collect(),
             canisters: Canisters::new(canisters, now),
-            canisters_directly_controlled_by_sns_root: BTreeSet::default(),
+            registry_canister_id,
             sns_root_canister: None,
             max_top_up_amount,
             min_interval,
@@ -133,6 +135,12 @@ pub struct Metrics {
     pub min_cycles_balance: Cycles,
     pub icp_burn_amount: Tokens,
     pub stable_memory_sizes: BTreeMap<u8, u64>,
-    pub ledger_canister: CanisterId,
-    pub cycles_minting_canister: CanisterId,
+    pub canister_ids: CanisterIds,
+}
+
+#[derive(CandidType, Serialize, Debug)]
+pub struct CanisterIds {
+    registry: CanisterId,
+    ledger: CanisterId,
+    cmc: CanisterId,
 }