+++ backend/canisters/notifications_index/c2c_client/src/lib.rs
@@ -1,0 +1,7 @@
+use canister_client::generate_c2c_call;
+use notifications_index_canister::*;
+
+// Queries
+
+// Updates
+generate_c2c_call!(add_notifications_canister);