+++ backend/canisters/registry/api/src/queries/mod.rs
@@ -1,2 +1,3 @@
 pub mod c2c_nervous_systems;
+pub mod subnets;
 pub mod updates;